refactor(silencer): split HOST/JOIN/TESTGAME into tick_hostjoin.cpp
Phase 4 mechanical move. Game::Tick's HOSTGAME, JOINGAME, and
TESTGAME case bodies become Game::TickHostGame/TickJoinGame/
TickTestGame in src/game/tick/tick_hostjoin.cpp. The dispatcher in
Game::Tick now calls into them. No behavior change.
